make use of the blacklist
the address (ip, port) is blacklisted instead of the identifier of the
peer. The code doesn't reveal that the address is blacklisted to keep
the offending peer in the dark.

more validation of inputs will be required. blacklisting could be done
at the protocol layer this allow to avoid scheduling a coroutine for
nothing.