known test failures
Depending on the network topology, Peer.get can fail because there is
no XOR that leads to a node that stored the value. To avoid that there
is the replication that is happening but set at 1 during the tests and
there is what is discribed in kademlia paper, page 2, paragraph 5:

   The Kademlia protocol ensures that every node knows at least one
   node in each of its subtrees, if that subtree contains a node.
   With this guarantee, any node can locate any other node by its ID.

This is not implemented in qadom as of yet.