workaround network split and bad network topologies
Add Peer.get_at, Peer.bag_at and Peer.namespace_at that take an extra
argument UID to reach the node called UID for the key peer is looking
for.  This is based on Peer._reach which might be optimize to avoid
fetch a value.

This fix the previous known failing test but a new one was discovered
that make test_namespace fail, see KNOWN-FAILING-SEEDS.  I am too lazy
to debug it right now...